Split tournament
Split tournament into init, ,models, views_main, views_feedback.
